Feat/auth page (#18)
* feat(frontend): Implement initial landing page (#4)

* chore(frontend): initial landing page commit

* chore(frontend): configure mock api data

introduced mockData.ts with sample trending albums and defined the Album interface. Added api.ts with a fetchReviews function to retrieve reviews from a mock API endpoint.

* chore(frontend): change header logo file

* chore(frontend): add hollow favicon SVG asset

Introduces a new SVG file 'arbust-favicon-hollow.svg' to the assets directory for use as a favicon in the frontend.

* style: remove logo final dot

* feat(frontend): add AuthPage with login and signup forms

Introduces a new AuthPage component with toggling login and signup forms, and adds a /login route to the app.

---------

Co-authored-by: Lucas Bonato <129971622+lucas-bonato@users.noreply.github.com>