Release v0.1.1
* Feat/auth page (#18)

* feat(frontend): Implement initial landing page (#4)

* chore(frontend): initial landing page commit

* chore(frontend): configure mock api data

introduced mockData.ts with sample trending albums and defined the Album interface. Added api.ts with a fetchReviews function to retrieve reviews from a mock API endpoint.

* chore(frontend): change header logo file

* chore(frontend): add hollow favicon SVG asset

Introduces a new SVG file 'arbust-favicon-hollow.svg' to the assets directory for use as a favicon in the frontend.

* style: remove logo final dot

* feat(frontend): add AuthPage with login and signup forms

Introduces a new AuthPage component with toggling login and signup forms, and adds a /login route to the app.

* Refactor/header n footer (#16)

* refactor: update layout to use Header and Footer components

Extracted header and footer into separate reusable components (Header.tsx and Footer.tsx) and updated BaseLayout to use them.

* feat(update): update page title in index.html

* Add reusable Input component (#17)

Introduces a new Input component with label support and customizable props.
Styled for consistent UI usage across the app.

* repo: add standard configuration (#19)

* repo: add pull request template

* repo: add CODEOWNERS file

* add workflow to block direct PRs to main branch

Introduces a GitHub Actions workflow that prevents PRs to the main branch unless they are from the 'develop' branch.